+++ src/main/java/com/fakng/fakngagrgtr/model/Company.java
@@ -1,6 +1,5 @@
-package com.fakng.fakngagrgtr.company;
+package com.fakng.fakngagrgtr.model;
 
-import com.fakng.fakngagrgtr.vacancy.Vacancy;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
