+++ src/main/java/com/fakng/fakngagrgtr/parser/HtmlParser.java
@@ -1,6 +1,5 @@
 package com.fakng.fakngagrgtr.parser;
 
-import com.fakng.fakngagrgtr.vacancy.Vacancy;
 import com.gargoylesoftware.htmlunit.WebClient;
 import com.gargoylesoftware.htmlunit.html.HtmlPage;
 import lombok.RequiredArgsConstructor;
@@ -8,7 +7,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
-import java.util.List;
 
 @Component
 @RequiredArgsConstructor
