+++ src/main/java/com/fakng/fakngagrgtr/repository/CompanyRepository.java
@@ -1,6 +1,6 @@
-package com.fakng.fakngagrgtr.company;
+package com.fakng.fakngagrgtr.repository;
 
-import com.fakng.fakngagrgtr.company.Company;
+import com.fakng.fakngagrgtr.model.Company;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
